fix: rename httperror package to errdecode and update references
