+++ internal/common/errdecode/error_format_test.go
@@ -1,10 +1,10 @@
-package httperror_test
+package errdecode_test
 
 import (
 	"strings"
 	"testing"
 
-	"github.com/kimdre/doco-cd/internal/common/httperror"
+	"github.com/kimdre/doco-cd/internal/common/errdecode"
 )
 
 func TestFormatErrorResponseBody_JSONObject(t *testing.T) {
@@ -12,7 +12,7 @@ func TestFormatErrorResponseBody_JSONObject(t *testing.T) {
 
 	body := []byte(`{"message":"\u041d\u0435\u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0430\u044f \u0432\u0435\u0442\u043a\u0430","typeName":"Microsoft.TeamFoundation.SourceControl.WebServer.InvalidRefNameException","errorCode":12345}`)
 
-	bodyText, structured := httperror.FormatErrorResponseBody(body)
+	bodyText, structured := errdecode.FormatErrorResponseBody(body)
 
 	if bodyText == "" {
 		t.Fatal("expected non-empty raw body text")
@@ -40,7 +40,7 @@ func TestFormatErrorResponseBody_NestedErrorObject(t *testing.T) {
 
 	body := []byte(`{"error":{"message":"nested failure","typeName":"SomeException","errorCode":42}}`)
 
-	_, structured := httperror.FormatErrorResponseBody(body)
+	_, structured := errdecode.FormatErrorResponseBody(body)
 
 	if !strings.Contains(structured, "message=nested failure") {
 		t.Errorf("expected message from nested error object, got: %s", structured)
@@ -56,7 +56,7 @@ func TestFormatErrorResponseBody_JSONString(t *testing.T) {
 
 	body := []byte(`"\u041d\u0435\u0434\u043e\u0441\u0442\u0443\u043f\u043d\u043e"`)
 
-	_, structured := httperror.FormatErrorResponseBody(body)
+	_, structured := errdecode.FormatErrorResponseBody(body)
 
 	if !strings.Contains(structured, "message=Недоступно") {
 		t.Errorf("expected decoded JSON string message, got: %s", structured)
@@ -68,7 +68,7 @@ func TestFormatErrorResponseBody_PlainText(t *testing.T) {
 
 	body := []byte("not-json body content")
 
-	bodyText, structured := httperror.FormatErrorResponseBody(body)
+	bodyText, structured := errdecode.FormatErrorResponseBody(body)
 
 	if bodyText != "not-json body content" {
 		t.Errorf("expected raw body text to be preserved, got: %s", bodyText)
@@ -82,7 +82,7 @@ func TestFormatErrorResponseBody_PlainText(t *testing.T) {
 func TestFormatErrorResponseBody_Empty(t *testing.T) {
 	t.Parallel()
 
-	bodyText, structured := httperror.FormatErrorResponseBody([]byte("   "))
+	bodyText, structured := errdecode.FormatErrorResponseBody([]byte("   "))
 
 	if bodyText != "" || structured != "" {
 		t.Errorf("expected empty results for blank body, got bodyText=%q structured=%q", bodyText, structured)
@@ -95,7 +95,7 @@ func TestDecodeEmbeddedJSON_ReplacesObjectInPlace(t *testing.T) {
 	text := "authentication required (or check https://dev.azure.com for error details)\n" +
 		`remote: {"message":"\u041d\u0435\u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0430\u044f \u0432\u0435\u0442\u043a\u0430","typeName":"Microsoft.TeamFoundation.SourceControl.WebServer.InvalidRefNameException","errorCode":12345}`
 
-	result := httperror.DecodeEmbeddedJSON(text)
+	result := errdecode.DecodeEmbeddedJSON(text)
 
 	if !strings.Contains(result, "Недопустимая ветка") {
 		t.Errorf("expected decoded message in result, got: %s", result)
@@ -127,7 +127,7 @@ func TestDecodeEmbeddedJSON_ReplacesBareEscapedString(t *testing.T) {
 
 	text := `authentication required: "\u041d\u0435\u0434\u043e\u0441\u0442\u0443\u043f\u043d\u043e"`
 
-	result := httperror.DecodeEmbeddedJSON(text)
+	result := errdecode.DecodeEmbeddedJSON(text)
 
 	if !strings.Contains(result, "Недоступно") {
 		t.Errorf("expected decoded text in result, got: %s", result)
@@ -143,7 +143,7 @@ func TestDecodeEmbeddedJSON_ReturnsOriginalWhenNothingToDecode(t *testing.T) {
 
 	text := "simple error message without JSON or escapes"
 
-	result := httperror.DecodeEmbeddedJSON(text)
+	result := errdecode.DecodeEmbeddedJSON(text)
 
 	if result != text {
 		t.Errorf("expected original text unchanged, got: %s", result)
@@ -157,7 +157,7 @@ func TestDecodeEmbeddedJSON_IgnoresPlainQuotedTextWithoutEscapes(t *testing.T) {
 	// likely unrelated plain text rather than a JSON-encoded localized message.
 	text := `some error: "just a quoted phrase"`
 
-	result := httperror.DecodeEmbeddedJSON(text)
+	result := errdecode.DecodeEmbeddedJSON(text)
 
 	if result != text {
 		t.Errorf("expected text with plain quotes to remain unchanged, got: %s", result)
